+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  - feat: `NesPressable` is only pressable when has listeners.
  - feat: adding `NesIcons.block`
  - feat: adding `NesIconButton.badge`
+ - feat: adding NesTheme.screenTransitionCursor
+ - chore: removing deprecated members from Color.
 
 # 0.20.0
  - feat: adding different frames to `NesDialog`s